+++ changelog.md
@@ -5,7 +5,8 @@
 ### Changed
 - List of coins in dropdown menu is alphabetical 
 ### Removed
-
+- Removed quick filter
+- Removed Graph for  Total 
 ## [0.1.0] 
 ### Added
 - Initial release